# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Muttley API
  description: >-
    API REST do Sistema Muttley para Gestão de Eventos Acadêmicos, Inscrições,
    Emissão e Validação de Certificados com Assinatura Visual e Gamificação por Medalhas.
    Desenvolvido para a FATEC Zona Leste.
  version: 0.1.0-alpha
  contact:
    name: Equipe Muttley - FATEC Zona Leste
    url: https://github.com/andrelamego/Backend-Muttley
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8083
    description: Servidor Local de Desenvolvimento
...
paths:
...
  /api/admin/disciplinas:
    get:
      tags:
        - Administração - Disciplinas
      summary: Listar disciplinas
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de disciplinas
    post:
      tags:
        - Administração - Disciplinas
      summary: Criar disciplina
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtualizacaoDisciplina'
      responses:
        '201':
          description: Criada com sucesso
  /api/admin/disciplinas/{id}:
    get:
      tags:
        - Administração - Disciplinas
      summary: Buscar disciplina por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Encontrada
    put:
      tags:
        - Administração - Disciplinas
      summary: Atualizar disciplina
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtualizacaoDisciplina'
      responses:
        '200':
          description: Atualizada
    delete:
      tags:
        - Administração - Disciplinas
      summary: Excluir disciplina
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Excluída
  /api/admin/enderecos:
    get:
      tags:
        - Administração - Endereços
      summary: Listar endereços
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de endereços
    post:
      tags:
        - Administração - Endereços
      summary: Criar endereço
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtualizacaoEndereco'
      responses:
        '201':
          description: Criado com sucesso
  /api/admin/enderecos/{id}:
    get:
      tags:
        - Administração - Endereços
      summary: Buscar endereço por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Encontrado
    put:
      tags:
        - Administração - Endereços
      summary: Atualizar endereço
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtualizacaoEndereco'
      responses:
        '200':
          description: Atualizado
    delete:
      tags:
        - Administração - Endereços
      summary: Excluir endereço
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Excluído
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Insira o token JWT retornado no endpoint /api/auth/login.
  schemas:
...
    AtualizacaoDisciplina:
      type: object
      required:
        - nome
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
          example: Banco de Dados Relacional
        sigla:
          type: string
          example: BDR
    AtualizacaoEndereco:
      type: object
      required:
        - logradouro
      properties:
        id:
          type: integer
          format: int64
        logradouro:
          type: string
          example: Av. Águia de Haia
        numero:
          type: string
          example: '2633'
        bairro:
          type: string
          example: Cidade A.E. Carvalho
        cidade:
          type: string
          example: São Paulo
        estado:
          type: string
          example: SP
        cep:
          type: string
          example: '03694000'
...